refactor: extract note tab widgets
