+++ lib/screens/home_screen.dart
@@ -1,18 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-import 'package:intl/intl.dart';
-import 'package:lottie/lottie.dart';
-import 'package:provider/provider.dart';
-import 'package:flutter/services.dart';
-import '../models/note.dart';
-import '../providers/note_provider.dart';
-import '../services/settings_service.dart';
-import '../widgets/add_note_dialog.dart';
-import '../widgets/notes_list.dart';
-import '../widgets/tag_filter_menu.dart';
+
+import '../widgets/notes_tab.dart';
 import 'chat_screen.dart';
 import 'note_list_for_day_screen.dart';
-import 'note_search_delegate.dart';
 import 'settings_screen.dart';
 import 'voice_to_note_screen.dart';
 
@@ -40,7 +31,10 @@ class _HomeScreenState extends State<HomeScreen> {
   void initState() {
     super.initState();
     _screens = [
-      const _NotesTab(),
+      NotesTab(
+        onThemeChanged: widget.onThemeChanged,
+        onFontScaleChanged: widget.onFontScaleChanged,
+      ),
       NoteListForDayScreen(date: DateTime.now()),
       const VoiceToNoteScreen(),
       const ChatScreen(initialMessage: ''),
@@ -60,10 +54,7 @@ class _HomeScreenState extends State<HomeScreen> {
         currentIndex: _currentIndex,
         onTap: (i) => setState(() => _currentIndex = i),
         items: [
-          const BottomNavigationBarItem(
-            icon: Icon(Icons.note),
-            label: 'Notes',
-          ),
+          const BottomNavigationBarItem(icon: Icon(Icons.note), label: 'Notes'),
           const BottomNavigationBarItem(
             icon: Icon(Icons.alarm),
             label: 'Reminders',
@@ -81,257 +72,6 @@ class _HomeScreenState extends State<HomeScreen> {
             label: l10n.settings,
           ),
         ],
-      ),
-    );
-  }
-}
-
-class _NotesTab extends StatefulWidget {
-  const _NotesTab();
-
-  @override
-  State<_NotesTab> createState() => _NotesTabState();
-}
-
-class _NotesTabState extends State<_NotesTab> {
-  String _mascotPath = 'assets/lottie/mascot.json';
-  final DateTime _today = DateTime.now();
-  String? _selectedTag;
-  static const _platform = MethodChannel('notes_reminder_app/actions');
-
-  @override
-  void initState() {
-    super.initState();
-    _loadMascot();
-    _platform.setMethodCallHandler((call) async {
-      if (call.method == 'voiceToNote') {
-        await Navigator.push(
-          context,
-          MaterialPageRoute(
-              builder: (_) => const VoiceToNoteScreen(autoStart: true)),
-        );
-      }
-    });
-  }
-
-  Future<void> _loadMascot() async {
-    _mascotPath = await SettingsService().loadMascotPath();
-    if (mounted) setState(() {});
-  }
-
-  void _addNote() {
-    showDialog(context: context, builder: (_) => const AddNoteDialog());
-  }
-
-  List<Note> _notesForDay(DateTime day, List<Note> notes) {
-    return notes
-        .where(
-          (n) =>
-              n.alarmTime != null &&
-              n.alarmTime!.year == day.year &&
-              n.alarmTime!.month == day.month &&
-              n.alarmTime!.day == day.day,
-        )
-        .toList();
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final provider = context.watch<NoteProvider>();
-    final notes = provider.notes;
-    final tags = notes.expand((n) => n.tags).toSet().toList();
-    final filteredNotes = _selectedTag == null
-        ? notes
-        : notes.where((n) => n.tags.contains(_selectedTag!)).toList();
-    final weekDays = List.generate(7, (i) => _today.add(Duration(days: i)));
-
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(AppLocalizations.of(context)!.appTitle),
-        actions: [
-          ValueListenableBuilder<SyncStatus>(
-            valueListenable: provider.syncStatus,
-            builder: (context, status, _) {
-              final l10n = AppLocalizations.of(context)!;
-              String text;
-              switch (status) {
-                case SyncStatus.syncing:
-                  text = l10n.syncStatusSyncing;
-                  break;
-                case SyncStatus.error:
-                  text = l10n.syncStatusError;
-                  break;
-                case SyncStatus.idle:
-                default:
-                  text = l10n.syncStatusIdle;
-              }
-              return Padding(
-                padding: const EdgeInsets.symmetric(horizontal: 8),
-                child: Center(child: Text(text)),
-              );
-            },
-          ),
-          TagFilterMenu(
-            tags: tags,
-            selectedTag: _selectedTag,
-            onSelected: (value) {
-              setState(() {
-                _selectedTag = value;
-              });
-            },
-          ),
-          IconButton(
-            icon: const Icon(Icons.search),
-            onPressed: () => showSearch(
-              context: context,
-              delegate: NoteSearchDelegate(context.read<NoteProvider>().notes),
-            ),
-          ),
-
-          IconButton(
-            icon: const Icon(Icons.mic),
-            onPressed: () async {
-              await Navigator.push(
-                context,
-                PageRouteBuilder(
-                  pageBuilder: (_, __, ___) => const VoiceToNoteScreen(),
-                  transitionsBuilder: (_, animation, __, child) {
-                    final offsetAnimation = Tween<Offset>(
-                      begin: const Offset(1, 0),
-                      end: Offset.zero,
-                    ).animate(animation);
-                    return FadeTransition(
-                      opacity: animation,
-                      child: SlideTransition(
-                        position: offsetAnimation,
-                        child: child,
-                      ),
-                    );
-                  },
-                ),
-              );
-            },
-          ),
-          IconButton(
-            icon: const Icon(Icons.settings),
-            tooltip: AppLocalizations.of(context)!.settingsTooltip,
-            onPressed: () async {
-              await Navigator.push(
-                context,
-                PageRouteBuilder(
-                  pageBuilder: (_, __, ___) => SettingsScreen(
-                    onThemeChanged: widget.onThemeChanged,
-                    onFontScaleChanged: widget.onFontScaleChanged,
-                  ),
-                  transitionsBuilder: (_, animation, __, child) {
-                    final offsetAnimation = Tween<Offset>(
-                      begin: const Offset(1, 0),
-                      end: Offset.zero,
-                    ).animate(animation);
-                    return FadeTransition(
-                      opacity: animation,
-                      child: SlideTransition(
-                        position: offsetAnimation,
-                        child: child,
-                      ),
-                    );
-                  },
-                ),
-              );
-              _loadMascot();
-            },
-          ),
-          PopupMenuButton<String>(
-            onSelected: (value) async {
-              if (value == 'backup') {
-                final ok = await context.read<NoteProvider>().backupNow();
-                if (!mounted) return;
-                final l10n = AppLocalizations.of(context)!;
-                if (ok) {
-                  ScaffoldMessenger.of(context).showSnackBar(
-                    SnackBar(content: Text(l10n.notesExported)),
-                  );
-                }
-              }
-            },
-            itemBuilder: (context) => [
-              PopupMenuItem(
-                value: 'backup',
-                child: Text(AppLocalizations.of(context)!.backupNow),
-              ),
-            ],
-          ),
-
-        ],
-      ),
-      body: Column(
-        children: [
-          const SizedBox(height: 8),
-          SizedBox(width: 140, height: 140, child: Lottie.asset(_mascotPath)),
-          const SizedBox(height: 8),
-          SizedBox(
-            height: 80,
-            child: ListView.builder(
-              scrollDirection: Axis.horizontal,
-              itemCount: weekDays.length,
-              itemBuilder: (context, i) {
-                final d = weekDays[i];
-                final hasNotes = _notesForDay(d, filteredNotes).isNotEmpty;
-                return GestureDetector(
-                  onTap: () {
-                    Navigator.push(
-                      context,
-                      PageRouteBuilder(
-                        pageBuilder: (_, __, ___) =>
-                            NoteListForDayScreen(date: d),
-                        transitionsBuilder: (_, animation, __, child) {
-                          final offsetAnimation = Tween<Offset>(
-                            begin: const Offset(1, 0),
-                            end: Offset.zero,
-                          ).animate(animation);
-                          return FadeTransition(
-                            opacity: animation,
-                            child: SlideTransition(
-                              position: offsetAnimation,
-                              child: child,
-                            ),
-                          );
-                        },
-                      ),
-                    );
-                  },
-                  child: Container(
-                    width: 60,
-                    margin: const EdgeInsets.all(4),
-                    decoration: BoxDecoration(
-                      color: hasNotes ? Colors.orange : Colors.white,
-                      border: Border.all(color: Colors.black),
-                      borderRadius: BorderRadius.circular(8),
-                    ),
-                    child: Column(
-                      mainAxisAlignment: MainAxisAlignment.center,
-                      children: [
-                        Text(
-                          DateFormat.E(
-                            Localizations.localeOf(context).toString(),
-                          ).format(d),
-                        ),
-                        Text('${d.day}'),
-                      ],
-                    ),
-                  ),
-                );
-              },
-            ),
-          ),
-          const SizedBox(height: 8),
-          Expanded(child: NotesList(notes: filteredNotes)),
-        ],
-      ),
-      floatingActionButton: FloatingActionButton(
-        onPressed: _addNote,
-        tooltip: AppLocalizations.of(context)!.addNoteTooltip,
-        child: const Icon(Icons.add),
       ),
     );
   }
